187_Semverを扱いやすく - Equalsとか - SemVerComparerを別ファイルに
