+++ Runtime/SemVerComparer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e78298e5bf6a77b41a23a44b512e556d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
